requirements.txt
django==4.2.7
django-compressor==4.4
django-libsass==0.9
django-meta==2.3.0
django-widget-tweaks==1.5.0
pillow==10.1.0
python-dotenv==1.0.0
django-tinymce==3.6.1
django-cleanup==8.0.0
django-taggit==5.0.1
django-admin-interface==0.28.5
whitenoise==6.6.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
requests==2.31.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-ckeditor==6.7.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
redis==5.0.1
django-prometheus==2.3.1
prometheus-client==0.17.1
